requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-apscheduler==0.5.2
django-environ==0.4.5
psycopg2==2.8.6
requests==2.25.1
django-extensions==3.1.3
pytest==6.2.4
